Update table of contents and documentation
Update table of contents and changed links to relative links. Also updated instructions to note that conda commands should be typed in Anaconda Prompt for Windows systems and system terminal for non-Windows systems, since conda installers may not write conda to the system path. 

Also note that we could control the trainees' environment better by specifying en environment.yaml file instead of a requirements.txt, since we can specify the Python version as well as the other packages. In fact, we should insist that they use a conda environment regardless of whether they have python on their machine (e.g. MacOS and Linux) for standardisation. Happy to create an issue for this.